Equip all components w/ a basic test examining the correct element type
This has been massive as we figured out quite a bit in this commit.
* First it is not possible to test styled-components just like that.
Instead you have to chain .dive() at the result of the shallow() output
in order to get in to pass.
* It is also not easy to test redux containers. Here we make a trick
and export the unconnected class only. i.e. import { Register } instead
of import Register; This way we are not testing any Redux specific code
but instead fully concentrate on React.
* It’s not easy to resolve webpack aliasing through babel-only /
non-webpack engines. For instance when running tests outside of
webpack, i.e. merely through jest than the aliased references cannot be
resolved. We empower babel via babel-jest to resolve the paths
according to what we specify in .babelrc -> plugins -> module-resolver.
The resolve power has been added to babel through the
‘babel-plugin-module-resolver’ package.