Stop using docker in development
Using docker in development is blocked by a bad experience when using
Hot Module Replacement. Currently, there is no proper way to sync file
changes on the host w/ the mounted volume inside the container. The
only viable option today is to use webpack-dev-server's watchMode. Yet
running the stack w/ these options requires way to much CPU and also
takes ages for code changes to be reflected in the browser. So while it
technically works, it's currently not suitable for real development
purposes.